requirements.txt
pytest~=6.2.5
eth-brownie>=1.11.0,<2.0.0
